requirements.txt
python-telegram-bot
httpx==0.27.0
apscheduler==3.10.4
asyncpg==0.29.0
pytz
